+++ MoveGenerator.h
@@ -12,7 +12,7 @@ public:
     inline MoveGenerator(Chess& pos) : ch(pos) {}
     Chess& ch;
     bool is_game_over(bool test);
-    uint8_t find_king(uint8_t color);
+    int find_king(int color);
     move* gen_moves(move (&moves)[120], bool test = false);
     void checks_exist(bool test);
     bool in_check = false;
@@ -23,9 +23,9 @@ public:
     static std::string move_san(Chess ch, move mv);
 private:
     void gen_pawn_moves(move (&pawn_moves)[120], bool test);
-    void gen_knight_piece_moves(move (&knight_moves)[120], uint8_t sq, bool test);
-    void gen_bishop_piece_moves(move (&bishop_moves)[120], uint8_t sq, bool test);
-    void gen_rook_piece_moves(move (&rook_moves)[120], uint8_t sq, bool test);
+    void gen_knight_piece_moves(move (&knight_moves)[120], int sq, bool test);
+    void gen_bishop_piece_moves(move (&bishop_moves)[120], int sq, bool test);
+    void gen_rook_piece_moves(move (&rook_moves)[120], int sq, bool test);
     void gen_king_piece_moves(move (&king_moves)[120], bool test);
     U64 find_pins(bool test);
     void check_method();
